requirements.txt
google-cloud-aiplatform
google-cloud-texttospeech
google-generativeai
moviepy
pillow
pydantic-settings
pydub
